+++ hutool-core/src/main/java/com/xiaoleilu/hutool/collection/CollUtil.java
@@ -724,7 +724,12 @@ public class CollUtil {
 	 */
 	public static <T> Collection<T> filter(Collection<T> collection, Editor<T> editor) {
 		Collection<T> collection2 = ObjectUtil.clone(collection);
-		collection2.clear();
+		try {
+			collection2.clear();
+		} catch (UnsupportedOperationException e) {
+			//克隆后的对象不支持清空，说明为不可变集合对象，使用默认的ArrayList保存结果
+			collection2 = new ArrayList<>();
+		}
 
 		T modified;
 		for (T t : collection) {
@@ -752,7 +757,12 @@ public class CollUtil {
 	 */
 	public static <T> Collection<T> filter(Collection<T> collection, Filter<T> filter) {
 		Collection<T> collection2 = ObjectUtil.clone(collection);
-		collection2.clear();
+		try {
+			collection2.clear();
+		} catch (UnsupportedOperationException e) {
+			//克隆后的对象不支持清空，说明为不可变集合对象，使用默认的ArrayList保存结果
+			collection2 = new ArrayList<>();
+		}
 
 		for (T t : collection) {
 			if (filter.accept(t)) {